Refactor package and method name. Test JDTBasedSpoonCompiler method. Add some doc in bash file.
