+++ chore/travis/travis-jdk9-other-seed.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
-# This script intends to be run on TravisCI
-# It executes compile and test goals
+# This script intends to be run on TravisCI only for commits on master branch
+# It compiles Spoon and launches tests with a seed for setting the order of compilation units
+# The purpose of this script is to check that all tests are passing even when the compilation units are not sorted in the same order
 if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then
     source /opt/jdk_switcher/jdk_switcher.sh
-    jdk_switcher use oraclejdk9 && SPOON_SEED_CU_COMPARATOR=$(( ( RANDOM % 10 )  + 1 )) mvn -Djava.src.version=1.9 test
+    jdk_switcher use oraclejdk9 && SPOON_SEED_CU_COMPARATOR=$(( ( RANDOM )  + 1 )) mvn -Djava.src.version=1.9 test
 fi
